# mcp-server/openapi.yaml
openapi: 3.1.0
info:
  title: PlotView API
  description: API to manage PlotView real estate projects and plots.
  version: 1.0.0
servers:
  - url: https://your-render-url.onrender.com
    description: Production Server (Replace with your actual URL)
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
security:
  - ApiKeyAuth: []
paths:
...
  /api/projects/{projectId}/plots:
    post:
      summary: Add a plot to a project
      operationId: addPlot
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                area:
                  type: number
                size:
                  type: string
                type:
                  type: string
                status:
                  type: string
                facing:
                  type: string
              required:
                - name
      responses:
        "200":
          description: Plot added successfully
...